requirements.txt
# ─── Core AI / ML ───────────────────────────────────────────────────────────
torch>=2.1.0
transformers>=4.40.0
accelerate>=0.29.0
peft>=0.10.0              # LoRA fine-tuning
trl>=0.8.6                # SFTTrainer for instruction tuning
datasets>=2.18.0          # HuggingFace datasets (CodeSearchNet)
sentence-transformers>=2.6.0  # Fallback embeddings
einops>=0.7.0             # Required by some transformer architectures
bitsandbytes>=0.43.0      # 4-bit/8-bit quantization (optional, saves GPU memory)

# ─── Embeddings + Vector DB ─────────────────────────────────────────────────
chromadb>=0.4.24
# CodeBERT loaded via transformers (microsoft/codebert-base)

# ─── RAG + LLM Orchestration ────────────────────────────────────────────────
langchain>=0.1.16
langchain-community>=0.0.34
langchain-core>=0.1.46

# ─── Code Parsing ───────────────────────────────────────────────────────────
tree-sitter>=0.21.3
tree-sitter-python>=0.21.0
tree-sitter-javascript>=0.21.0
tree-sitter-typescript>=0.21.0
tree-sitter-java>=0.21.0
gitpython>=3.1.43          # Git history parsing for changelog

# ─── API Server (AI microservice) ───────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
openai>=1.30.0             # OpenAI SDK — used for DO Gradient AI (OpenAI-compatible API)
httpx>=0.27.0              # Async HTTP client
redis>=5.0.3               # Caching layer
pydantic>=2.6.4

# ─── Storage + Utilities ────────────────────────────────────────────────────
boto3>=1.34.0              # DigitalOcean Spaces (S3-compatible)
python-dotenv>=1.0.1
pyyaml>=6.0.1
tqdm>=4.66.2
numpy>=1.26.4
